유저 데이터 가져오는 로직을 cutsom hook으로 분리
